+++ genium/src/test/resources/smoke/errors/output/android/jni/com_example_smoke_Errors.h
@@ -2,24 +2,19 @@
  *
  * Automatically generated. Do not modify. Your changes will be lost.
  */
-
 #pragma once
-
 #include <jni.h>
-
 #ifdef __cplusplus
 extern "C" {
 #endif
-
 JNIEXPORT void JNICALL
 Java_com_example_smoke_Errors_methodWithErrors(JNIEnv* _jenv, jobject _jinstance);
-
 JNIEXPORT void JNICALL
 Java_com_example_smoke_Errors_methodWithExternalErrors(JNIEnv* _jenv, jobject _jinstance);
-
 JNIEXPORT jstring JNICALL
 Java_com_example_smoke_Errors_methodWithErrorsAndReturnValue(JNIEnv* _jenv, jobject _jinstance);
-
+JNIEXPORT void JNICALL
+Java_com_example_smoke_Errors_methodWithAliasedError(JNIEnv* _jenv, jobject _jinstance);
 #ifdef __cplusplus
 }
 #endif