+++ genium/src/test/resources/smoke/errors/output/cbridge/include/smoke/cbridge_Errors.h
@@ -24,6 +24,7 @@ _GENIUM_C_EXPORT _baseRef smoke_Errors_copy_handle(_baseRef handle);
 _GENIUM_C_EXPORT smoke_Errors_InternalError smoke_Errors_methodWithErrors();
 _GENIUM_C_EXPORT smoke_Errors_ExternalErrors smoke_Errors_methodWithExternalErrors();
 _GENIUM_C_EXPORT smoke_Errors_methodWithErrorsAndReturnValue_result smoke_Errors_methodWithErrorsAndReturnValue();
+_GENIUM_C_EXPORT smoke_Errors_InternalError smoke_Errors_methodWithAliasedError();
 #ifdef __cplusplus
 }
 #endif