+++ genium/src/test/resources/smoke/errors/output/cbridge/src/smoke/cbridge_Errors.cpp
@@ -32,3 +32,6 @@ smoke_Errors_methodWithErrorsAndReturnValue_result smoke_Errors_methodWithErrors
         return {false, .error_code = static_cast< smoke_Errors_InternalError >(RESULT.error().value())};
     }
 }
+smoke_Errors_InternalError smoke_Errors_methodWithAliasedError() {
+    return ::smoke::Errors::method_with_aliased_error().value();
+}