+++ genium/src/test/resources/smoke/errors/output/cpp/include/smoke/Errors.h
@@ -22,10 +22,12 @@ enum class InternalError {
     ERROR_NONE,
     ERROR_FATAL
 };
+using Boom = ::std::error_code;
 public:
 static ::std::error_code method_with_errors(  );
 static ::std::error_code method_with_external_errors(  );
 static ::genium::Return< ::std::string, ::std::error_code > method_with_errors_and_return_value(  );
+static ::std::error_code method_with_aliased_error(  );
 };
 _GENIUM_CPP_EXPORT ::std::error_code make_error_code( ::smoke::Errors::InternalError value ) noexcept;
 }